// src/app/interfaces/IPositionPattern.interface.ts
  export interface IPositionPattern{
	id:     		 number;
	creation:       string;
	modification:   string;
	name:           string;
	description:    string;
	status:			string;
	note:			string;
	active:			number;
	deleted:		number;
  }

  export interface IPositionHighPattern{
	id:     		 number;
	creation:       string;
	modification:   string;
	name:           string;
	description:    string;
	status:			string;
	note:			string;
	active:			number;
	deleted:		number;
  }